Separated the Redux for Steem actions
All the Steem blockchain based Redux actions and reducers were in one file. I separated them into their own reducers and actions.
