+++ client/src/components/pages/Steem/Posts.js
@@ -10,7 +10,9 @@ import ErrorLabel from '../../ErrorLabel/ErrorLabel';
 import ErrorBoundary from '../../ErrorBoundary/ErrorBoundary';
 import Loading from '../../Loading/Loading';
 import FilterPosts from './FilterPosts';
-import * as contentActions from '../../../actions/steemContentActions';
+import { getSummaryContent } from '../../../actions/summaryPostActions';
+import * as addPostActions from '../../../actions/addPostActions';
+import { upvotePost } from '../../../actions/upvoteActions';
 
 /**
  *  Gets the Steem blockchain content and displays a list of post
@@ -247,20 +249,27 @@ class Posts extends Component {
 const mapStateToProps = state => {
   const {
     auth: {
-      user, csrf
+      user,
+      csrf,
     },
-    steemContent: {
-      posts,
+    summaryPost: {
       isFetchingSummary,
       prevPage,
       noMore,
+      posts,
+    },
+    userGroups: {
       groups,
+    },
+    addPost: {
       postExists,
       addPostLoading,
       modalOpenAddPost,
       selectedGroup,
+    },
+    upvote: {
       upvotePayload,
-    }
+    },
   } = state;
 
   return {
@@ -282,22 +291,22 @@ const mapStateToProps = state => {
 const mapDispatchToProps = (dispatch) => (
   {
     getContent: (selectedFilter, query, nextPost, page) => (
-      dispatch(contentActions.getSummaryContent(selectedFilter, query, nextPost, page))
+      dispatch(getSummaryContent(selectedFilter, query, nextPost, page))
     ),
     showModal: (e, type, data) => (
-      dispatch(contentActions.showModal(e, type, data))
+      dispatch(addPostActions.showModal(e, type, data))
     ),
     handleModalClickAddPost: (e) => (
-      dispatch(contentActions.handleModalClickAddPost(e))
+      dispatch(addPostActions.handleModalClickAddPost(e))
     ),
     onModalCloseAddPost: () => (
-      dispatch(contentActions.onModalCloseAddPost())
+      dispatch(addPostActions.onModalCloseAddPost())
     ),
     handleGroupSelect: (value) => (
-      dispatch(contentActions.handleGroupSelect(value))
+      dispatch(addPostActions.handleGroupSelect(value))
     ),
     handleUpvote: (author, permlink, weight) => (
-      dispatch(contentActions.upvotePost(author, permlink, weight))
+      dispatch(upvotePost(author, permlink, weight))
     ),
   }
 );
