+++ client/src/reducers/index.js
@@ -2,7 +2,13 @@ import { combineReducers } from 'redux';
 
 import { recentActivity, selected } from './recentPostsReducer';
 import { auth } from './authReducer';
-import { steemContent } from './steemContentReducer';
+import { summaryPost } from './summaryPostReducer';
+import { userGroups } from './userGroupsReducer';
+import { detailsPost } from './detailsPostReducer';
+import { addPost } from './addPostReducer';
+import { comments } from './commentsReducer';
+import { upvote } from './upvoteReducer';
+import { sendComment } from './sendCommentReducer';
 
 /**
  *  Combine imported reducers for the application.
@@ -11,7 +17,13 @@ const rootReducer = combineReducers({
   recentActivity,
   selected,
   auth,
-  steemContent,
+  summaryPost,
+  userGroups,
+  detailsPost,
+  addPost,
+  comments,
+  upvote,
+  sendComment,
 });
 
 export default rootReducer;
